Move OrganisationSeedData back to Data project
